+++ inst/srcjs/shinyjs-default-funcs.js
@@ -754,11 +754,10 @@ shinyjs = function() {
           // file inputs need to be reset manually since shiny doesn't have an
           // update function for them
           if (type == "File") {
-            id = "#" + id;
-            $(id).val('');
-            $(id + "_progress").css("visibility", "hidden");
-            $(id + "_progress").find(".progress-bar").css("width", "0");
-            $(id).closest(".input-group").find("input[type='text']").val('');
+            _jqid(id).val('');
+            _jqid(id + "_progress").css("visibility", "hidden");
+            _jqid(id + "_progress").find(".progress-bar").css("width", "0");
+            _jqid(id).closest(".input-group").find("input[type='text']").val('');
           } else {
             messages[id] = { 'type' : type, 'value' : value };
           }
@@ -769,8 +768,8 @@ shinyjs = function() {
       Shiny.onInputChange(params.shinyInputId, messages);
     },
 
-   // run an R function after an asynchronous delay
-   delay : function(params) {
+    // run an R function after an asynchronous delay
+    delay : function(params) {
       var defaultParams = {
         ms : null,
         shinyInputId : null
@@ -781,7 +780,7 @@ shinyjs = function() {
       setTimeout(function() {
         Shiny.onInputChange(params.shinyInputId, params.ms);
       }, params.ms);
-   },
+    },
 
     // click on a button
     click : function(params) {