+++ R/jsFunc-html.R
@@ -8,7 +8,7 @@
 #' @param html The HTML/text to place inside the element. Can be either simple
 #' plain text or valid HTML code.
 #' @param add If \code{TRUE}, then append \code{html} to the contents of the element;
-#' otherwise overwrite it. (default: \code{FALSE})
+#' otherwise overwrite it.
 #' @param selector JQuery selector of the elements to target. Ignored if the \code{id}
 #' argument is given.
 #' @seealso \code{\link[shinyjs]{useShinyjs}},
@@ -43,8 +43,8 @@
 #'   html(id = "element", add = TRUE, html = input$btn)
 #' }
 #' @export
-html <- function(id, html, add, selector) {
+html <- function(id = NULL, html = NULL, add = FALSE, selector = NULL) {
   fxn <- "html"
-  params <- as.list(match.call())[-1]
+  params <- list(id = id, html = html, add = add, selector = selector)
   jsFuncHelper(fxn, params)
 }