+++ NEWS.md
@@ -1,6 +1,7 @@
 # shinyjs 0.9.0.9xxx
 
-- fixed bug where using `disabled()` on certain non-standard inputs (such as `shinyFiles`) had strange behaviour #120
+- Bug fix: `reset()` wasn't working on checkboxes/select inputs that had comma in the value (#130)
+- fixed bug where using `disabled()` on certain non-standard inputs (such as `shinyFiles`) had strange behaviour (#120)
 - fixed bug where logjs()+showLog() wasn't working when app first runs
 
 # shinyjs 0.9